Made all the on_messages be in one file
